requirements.txt
python-telegram-bot==21.6
feedparser==6.0.11
httpx==0.27.2
